refactor: move cobra/cmd/* to cobra/
