+++ cobra/testdata/LICENSE.golden
