+++ cobra/testdata/main.go.golden
@@ -15,7 +15,7 @@ limitations under the License.
 */
 package main
 
-import "github.com/spf13/cobra/cobra/cmd/cmd"
+import "github.com/spf13/cobra/cobra/cmd"
 
 func main() {
 	cmd.Execute()
