+++ cobra/testdata/test.go.golden
